+++ routes/~_dashboard/~notifications/helpers.tsx
@@ -1,0 +1,8 @@
+import { createColumnHelper } from '@tanstack/table-core'
+import { UserMessage } from '@/modules/firebase/models'
+
+export const columnHelper = createColumnHelper<UserMessage>()
+
+export const columnIds = {
+  isRead: 'isRead',
+}
